allow only admin to update and delete

+++ src/components/DeleteEntry.js
@@ -1,23 +1,31 @@
 import React, { useContext } from "react";
 import { UpdateContext } from "../context/UpdateContext";
+import { AuthContext } from "../context/AuthContext";
 
 const DeleteEntry = ({ entryType, entry, onClose }) => {
     const { setWordUpdate } = useContext(UpdateContext);
+    const { userRole } = useContext(AuthContext);
+
     const entryId = entry._id;
+
     const handleSubmit = async () => {
-        const url = `${process.env.REACT_APP_SERVER_URL}/${entryType}/${entryId}`;
-        const res = await fetch(url, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json",
-            },
-        });
-        const data = await res.json();
-        if (res.status === 200) {
-            onClose();
-            setWordUpdate((prev) => !prev);
+        if (userRole === "admin") {
+            const url = `${process.env.REACT_APP_SERVER_URL}/${entryType}/${entryId}`;
+            const res = await fetch(url, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+            });
+            const data = await res.json();
+            if (res.status === 200) {
+                onClose();
+                setWordUpdate((prev) => !prev);
+            } else {
+                window.alert(data.error);
+            }
         } else {
-            window.alert(data.error);
+            window.alert("Only admin can delete");
         }
     };
